[REF] web,*: remove some legacy test helpers
Part of task~3439226
